Reconstruct with use context
